requirements.txt
aiogram==3.13.1
python-dotenv==1.0.1
playwright==1.49.0
pydantic==2.9.2

